initial implementation with MariaDB
changes in the structure and started the implementation using MariaDB
