+++ src/templates/index.html
@@ -1,0 +1,30 @@
+{% extends "base.html" %}
+
+{% block main %}
+    <div>
+        <h1>API Description</h1>
+        <p>
+            <a>API to get information of one specific User:</a>
+            <br>
+            <a>additional parameters in Format ?legacyAddress=XYZ&signature=SIG&IP=xxx.xxx needed</a>
+            <br>
+            <a href="/api/v2/userInformation?legacyAddress=XYZ&amp;signature=SIG&amp;ip=xxx.xxx">/api/v2/userInformation</a>
+            <br>
+        </p>
+        <p>
+            <a>API to get all Registrations of one specific User</a>
+            <br>
+            <a>additional parameter in Format ?legacyAddress=XYZ&IP=xxx.xxx needed</a>
+            <br>
+            <a href="/api/v2/registrations?legacyAddress=XYZ">/api/v2/registrations</a>
+            <br>
+        </p>
+        <p>
+            <a>API to post new Registration of one specific User:</a>
+            <br>
+            <a>additional parameters in Format ?legacyAddress=XYZ&signature=SIG needed</a>
+            <br>
+            <a href="/api/v2/addRegistration?legacyAddress=XYZ&amp;signature=SIG">/api/v2/addRegistration</a>
+            <br>
+        </p>
+{% endblock %}
